Create SearchField component
