+++ src/components/SearchField/index.ts
@@ -1,0 +1,1 @@
+export { default as SearchField } from './SearchField';
